+++ spark/src/test/java/org/apache/iceberg/actions/TestRewriteDataFilesAction.java
@@ -19,22 +19,10 @@
 
 package org.apache.iceberg.actions;
 
-import com.esotericsoftware.kryo.Kryo;
-import com.esotericsoftware.kryo.io.Input;
-import com.esotericsoftware.kryo.io.Output;
-import java.io.ByteArrayInputStream;
-import java.io.ByteArrayOutputStream;
 import java.io.File;
-import java.io.FileInputStream;
-import java.io.FileOutputStream;
-import java.io.ObjectInputStream;
-import java.io.ObjectOutputStream;
-import java.util.Comparator;
 import java.util.List;
 import java.util.Map;
-import java.util.stream.Collectors;
 import org.apache.hadoop.conf.Configuration;
-import org.apache.iceberg.BaseCombinedScanTask;
 import org.apache.iceberg.DataFile;
 import org.apache.iceberg.FileScanTask;
 import org.apache.iceberg.PartitionSpec;
@@ -49,8 +37,6 @@ import org.apache.iceberg.relocated.com.google.common.collect.Maps;
 import org.apache.iceberg.spark.SparkTestBase;
 import org.apache.iceberg.spark.source.ThreeColumnRecord;
 import org.apache.iceberg.types.Types;
-import org.apache.spark.SparkConf;
-import org.apache.spark.serializer.KryoSerializer;
 import org.apache.spark.sql.Dataset;
 import org.apache.spark.sql.Encoders;
 import org.apache.spark.sql.Row;
@@ -323,128 +309,6 @@ public abstract class TestRewriteDataFilesAction extends SparkTestBase {
         .collectAsList();
 
     Assert.assertEquals("Rows must match", records, actualRecords);
-  }
-
-  @Test
-  public void testBaseCombinedScanTaskKryoSerialization() throws Exception {
-    BaseCombinedScanTask scanTask = prepareBaseCombinedScanTaskForSerDeTest();
-
-    File data = temp.newFile();
-    Assert.assertTrue(data.delete());
-    Kryo kryo = new KryoSerializer(new SparkConf()).newKryo();
-
-    try (Output out = new Output(new FileOutputStream(data))) {
-      kryo.writeClassAndObject(out, scanTask);
-    }
-
-    try (Input in = new Input(new FileInputStream(data))) {
-      Object obj = kryo.readClassAndObject(in);
-      Assert.assertTrue("Should be a BaseCombinedScanTask", obj instanceof BaseCombinedScanTask);
-      checkBaseCombinedScanTask(scanTask, (BaseCombinedScanTask) obj);
-    }
-  }
-
-  @Test
-  public void testBaseCombinedScanTaskJavaSerialization() throws Exception {
-    BaseCombinedScanTask scanTask = prepareBaseCombinedScanTaskForSerDeTest();
-
-    ByteArrayOutputStream bytes = new ByteArrayOutputStream();
-    try (ObjectOutputStream out = new ObjectOutputStream(bytes)) {
-      out.writeObject(scanTask);
-    }
-
-    try (ObjectInputStream in = new ObjectInputStream(new ByteArrayInputStream(bytes.toByteArray()))) {
-      Object obj = in.readObject();
-      Assert.assertTrue("Should be a BaseCombinedScanTask", obj instanceof BaseCombinedScanTask);
-      checkBaseCombinedScanTask(scanTask, (BaseCombinedScanTask) obj);
-    }
-  }
-
-  private BaseCombinedScanTask prepareBaseCombinedScanTaskForSerDeTest() {
-    PartitionSpec spec = PartitionSpec.unpartitioned();
-    Map<String, String> options = Maps.newHashMap();
-    Table table = TABLES.create(SCHEMA, spec, options, tableLocation);
-
-    List<ThreeColumnRecord> records1 = Lists.newArrayList(
-            new ThreeColumnRecord(1, null, "AAAA"),
-            new ThreeColumnRecord(1, "BBBBBBBBBB", "BBBB")
-    );
-    writeRecords(records1);
-
-    List<ThreeColumnRecord> records2 = Lists.newArrayList(
-            new ThreeColumnRecord(2, "CCCCCCCCCC", "CCCC"),
-            new ThreeColumnRecord(2, "DDDDDDDDDD", "DDDD")
-    );
-    writeRecords(records2);
-
-    table.refresh();
-
-    CloseableIterable<FileScanTask> tasks = table.newScan().planFiles();
-    return new BaseCombinedScanTask(Lists.newArrayList(tasks));
-  }
-
-  private void checkBaseCombinedScanTask(BaseCombinedScanTask expected, BaseCombinedScanTask actual) {
-    List<FileScanTask> expectedTasks = getFileScanTasksInFilePathOrder(expected);
-    List<FileScanTask> actualTasks = getFileScanTasksInFilePathOrder(actual);
-
-    Assert.assertEquals("The number of file scan tasks should match",
-            expectedTasks.size(), actualTasks.size());
-
-    for (int i = 0; i < expectedTasks.size(); i++) {
-      FileScanTask expectedTask = expectedTasks.get(i);
-      FileScanTask actualTask = actualTasks.get(i);
-      checkFileScanTask(expectedTask, actualTask);
-    }
-  }
-
-  private List<FileScanTask> getFileScanTasksInFilePathOrder(BaseCombinedScanTask task) {
-    return task.files().stream()
-            // use file path + start position to differentiate the tasks
-            .sorted(Comparator.comparing(o -> o.file().path().toString() + "##" + o.start()))
-            .collect(Collectors.toList());
-  }
-
-  private void checkFileScanTask(FileScanTask expected, FileScanTask actual) {
-    checkDataFile(expected.file(), actual.file());
-
-    // PartitionSpec implements its own equals method
-    Assert.assertEquals("PartitionSpec doesn't match", expected.spec(), actual.spec());
-
-    Assert.assertEquals("starting position doesn't match", expected.start(), actual.start());
-
-    Assert.assertEquals("the number of bytes to scan doesn't match", expected.start(), actual.start());
-
-    // simplify comparison on residual expression via comparing toString
-    Assert.assertEquals("Residual expression doesn't match",
-            expected.residual().toString(), actual.residual().toString());
-  }
-
-  // TODO: this is a copy of TestDataFileSerialization.checkDataFile, deduplicate
-  private void checkDataFile(DataFile expected, DataFile actual) {
-    Assert.assertEquals("Should match the serialized record path",
-            expected.path(), actual.path());
-    Assert.assertEquals("Should match the serialized record format",
-            expected.format(), actual.format());
-    Assert.assertEquals("Should match the serialized record partition",
-            expected.partition().get(0, Object.class), actual.partition().get(0, Object.class));
-    Assert.assertEquals("Should match the serialized record count",
-            expected.recordCount(), actual.recordCount());
-    Assert.assertEquals("Should match the serialized record size",
-            expected.fileSizeInBytes(), actual.fileSizeInBytes());
-    Assert.assertEquals("Should match the serialized record value counts",
-            expected.valueCounts(), actual.valueCounts());
-    Assert.assertEquals("Should match the serialized record null value counts",
-            expected.nullValueCounts(), actual.nullValueCounts());
-    Assert.assertEquals("Should match the serialized record lower bounds",
-            expected.lowerBounds(), actual.lowerBounds());
-    Assert.assertEquals("Should match the serialized record upper bounds",
-            expected.upperBounds(), actual.upperBounds());
-    Assert.assertEquals("Should match the serialized record key metadata",
-            expected.keyMetadata(), actual.keyMetadata());
-    Assert.assertEquals("Should match the serialized record offsets",
-            expected.splitOffsets(), actual.splitOffsets());
-    Assert.assertEquals("Should match the serialized record offsets",
-            expected.keyMetadata(), actual.keyMetadata());
   }
 
   private void writeRecords(List<ThreeColumnRecord> records) {
